Tweak FineCommandParserTest to handle new compound predicate
